requirements.txt
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1